feat(config): include a default config file
This clears the internal structures:
defaultConfig, defaultKeys, and defaultTheme.

Then includes a config.toml and then parses it, instead of duplicating
data or letting a possible future config file and actual in-program
default values not being equal.

This also forces any future configs to have proper documentation
and an example attached to them that non-developers can read.

Fixes #459