Add mention support
You need to press Tab to load the user into the MemberState if the user is not
in the currently loaded messages.

This is not similar to how Discord does it, because when Discord differentiate
between a literal `@SomeGuy` and a mention of SomeGuy.

If you want to include a literal `@SomeUser` without actually pinging them,
you need to wrap them in code inlines.

There could be a way of doing it if we can encode data *alongside* text when
typing (e.g., encode the mention interally and display it to the user as
colored text and when editing the text that color is encoded
as `<@SomeUser:12345678>` when editing with an external editor and when
pasting/copying text??)